[feat] Support alternate port without requiring Docker port mapping
On CircleCI, and I assume other environments, the use of the [Docker `-p`
option](https://docs.docker.com/engine/reference/run/#expose-incoming-ports_) to map
ports [is not
available](https://circleci.com/docs/2.0/executor-types/#using-multiple-docker-images)
for security reasons.  This configures nginx to listen on and Docker to expose a random,
high port that is unlikely to conflict with anything else.